+++ k-distribution/src/main/scripts/lib/pyk/ktool.py
@@ -5,7 +5,7 @@ from   pathlib    import Path
 import subprocess
 
 from .cli_utils import gen_file_timestamp, notif
-from .util import getAppliedAxiomList, strHash
+from .utils import getAppliedAxiomList, strHash
 from .kastManip import *
 
 class KPrint:
